pyproject.toml
[tool.poetry]
name = "hacker-the-future-api"
version = "0.1.0"
description = ""
authors = ["Enmanuel Silva <silvacompte@gmail.com>"]

[tool.poetry.dependencies]
alembic = "^1.9.2"
fastapi = "^0.89.1"
psycopg2-binary = "^2.9.5"
pydantic = {extras = ["email"], version = "^1.10.5"}
python = "^3.9"
python-dotenv = "^0.21.1"
sqladmin = "^0.8.0"
sqlmodel = "^0.0.8"
uvicorn = {extras = ["standard"], version = "^0.20.0"}

[tool.poetry.dev-dependencies]
black = {version = "^23.1.0", allow-prereleases = true}
commitizen = "^2.40.0"
pre-commit = "^3.0.1"
pytest = "^7.2.2"
pycln = "^2.1.3"
pytest-emoji = "^0.2.0"
pytest-md-report = "^0.3.0"
pytest-xdist = "^3.2.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.black]
line-length = 79

[tool.pycln]
all = true

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
    "src/__version__.py",
    "pyproject.toml:version"
]
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]
